Recreate via as the new name for typed_via
PR #538 deleted `unifex::via()`, which was an untyped _Sender_
algorithm, and changed all uses to use `unifex::typed_via()` instead.
This diff recreates the `via.hpp` header; it now defines the "typed via"
algorithm under the name `unifex::via()`. At the same time, this diff
changes `unifex::typed_via` to be a deprecated alias for the `via`
algorithm.
